requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
twilio>=8.0.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-allauth>=0.54.0
python-decouple>=3.8
reportlab>=4.0.0
django-widget-tweaks>=1.5.0
django-cors-headers>=4.2.0
whitenoise>=6.5.0
gunicorn>=21.2.0
google-generativeai>=0.3.0 
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.20.0
pillow>=9.0.0
django>=4.2.0